requirements.txt
websocket-client
qpython
pymysql
numpy
pyzmq
